removed SR_READ feature. it was confusing in the implementation, and no-one will ever use it.
Added optimization to skip files in /proc/*/fd/ that don't start with '/'
such files are sockets and pipes, and we cannot post them.

Added code to detect and communicate failure to connect.
(addresses some seg faults, when connection assumed but context is bad pointer)

messy patch, should not do this...