remove file_copy_range
In #111, there are segmentation faults on ubuntu 24.04. I do not know
why. This routine accepts file descriptors... and those descriptors
need to be subsequently closed, so the file will be published when the
close happens anyways. There is no need for this entry point.

When I remove this routine, stuff stops crashing. One possibility is
that calculating the checksum involves reading the file again, possibly
opening and re-closing it. just a guess. I hope I am not just masking
the problem by removing a simple cause, but the change is good anyways.